// apps/api/src/modules/workspace/enums/workspace.status-code.enum.ts
export enum ENUM_WORKSPACE_STATUS_CODE_ERROR {
    NOT_FOUND = 5200,
    NAME_EXIST = 5201,
    USER_NOT_OWNER = 5202,
    CREATE_FAILED = 5203,
    UPDATE_FAILED = 5204,
    DELETE_FORBIDDEN = 5205,
    MEMBER_EXIST = 5206,
    MEMBER_NOT_EXIST = 5207,
    MEMBER_NOT_FOUND = 5208,
    WORKSPACE_EXIST = 5209,
    INVITATION_LINK_EXPIRED = 5210,
    INVITATION_LINK_INVALID = 5211,
    INVALID_INVITATION_CODE = 5212,
}